storage: move CatalogSnapshot and IndexInfo into catalog_cache.rs
Both types and the extract_columns_from_index_sql helper were embedded in
btree.rs alongside thousands of lines of cursor and I/O logic. Separating
them into catalog_cache.rs makes each file's concern clear:

- catalog_cache.rs: catalog metadata types and column parsing
- btree.rs: B-tree structure, cursor navigation, catalog read/write ops
