+++ src/storage.rs
@@ -1,3 +1,4 @@
+mod catalog_cache;
 mod cell;
 mod cell_reader;
 mod error;
@@ -21,7 +22,7 @@ pub use btree::encode_integer_key;
 pub use btree::encode_u64_key;
 pub use btree::BTree;
 pub use btree::CursorHandle;
-pub use btree::IndexInfo;
+pub use catalog_cache::IndexInfo;
 pub use cell_reader::CellReader;
 pub use node_page_store::NodePageStore;
 pub use error::Error as StorageError;
